More Fixes

+++ TvMazeClient/Models/Externals.cs
@@ -5,10 +5,10 @@ namespace TvMazeClient.Models;
 public partial class Externals
 {
     [JsonPropertyName("tvrage")]
-    public long Tvrage { get; set; }
+    public int Tvrage { get; set; }
 
     [JsonPropertyName("thetvdb")]
-    public long Thetvdb { get; set; }
+    public int Thetvdb { get; set; }
 
     [JsonPropertyName("imdb")]
     public string Imdb { get; set; }
